Split bb into files
