+++ qbee/util.py
@@ -1,33 +1,24 @@
 import sympy as sp
-
+import numpy as np
 from time import time
 from tqdm import tqdm
-from functools import partial
-from typing import List, Optional, Iterable
+from typing import Iterable, Collection
 from sympy.polys.monomials import monomial_deg
 from itertools import combinations
+from functools import reduce
+from operator import add
 
 
-def posify_monomial(monomial: sp.Monomial):
-    # TODO: Odd, it was useful with negative substitutions. Has to check this out.
-    return monomial if '-' not in str(monomial) else -monomial
+def timed(func):
+    def wrapper(*args, **kwargs):
+        start_time = time()
+        res = func(*args, **kwargs)
+        end_time = time()
+        print()
+        print(f"Elapsed time: {np.round(end_time - start_time, 3)}s.")
+        return res
 
-
-def polynomial_subs(poly: sp.Poly, old: sp.Poly, new: sp.Poly) -> sp.Expr:
-    monomials = sp.Add.make_args(poly)
-    return sp.Add(*map(partial(monomial_subs, old=old, new=new), monomials))
-
-
-def monomial_subs(monomial: sp.Expr, old: sp.Expr, new: sp.Expr) -> sp.Expr:
-    quotient, rem = sp.polys.div(monomial, old)
-    if rem == 0:
-        return quotient * new
-    else:
-        return monomial
-
-
-def is_monomial_divisor(numerator: sp.Expr, denominator: sp.Expr) -> bool:
-    return sp.gcd(numerator.as_expr(), denominator.as_expr()) == denominator.as_expr()
+    return wrapper
 
 
 def reset_progress_bar(pbar: tqdm, value):
@@ -42,6 +33,18 @@ def refresh_and_close_progress_bars(*pbars: tqdm):
         bar.close()
 
 
+def get_decompositions(monomial):
+    if len(monomial) == 0:
+        return {(tuple(), tuple())}
+    result = set()
+    prev_result = get_decompositions(tuple(monomial[:-1]))
+    for r in prev_result:
+        for i in range(monomial[-1] + 1):
+            a, b = tuple(list(r[0]) + [i]), tuple(list(r[1]) + [monomial[-1] - i])
+            result.add((min(a, b), max(a, b)))
+    return result
+
+
 def symbol_from_derivative(derivative: sp.Symbol) -> sp.Symbol:
     return sp.Symbol(str(derivative).replace(r"\dot ", '', 1))
 
@@ -54,33 +57,14 @@ def monomial_to_poly(monom: sp.Monomial) -> sp.Poly:
     return sp.Poly(sp.prod([gen ** e for gen, e in zip(monom.gens, monom.exponents)]), monom.gens)
 
 
-def unify_monom_to_gens(in_monom: sp.Monomial, gens: List[sp.Symbol]):
-    res = list()
-    for gen in gens:
-        gen_index = index_or_None(in_monom.gens, gen)
-        if gen_index is not None:
-            res.append(in_monom.exponents[gen_index])
-        else:
-            res.append(0)
-    return sp.Monomial(res, gens)
-
-
-def index_or_None(it: Iterable, elem) -> Optional[int]:
-    for i, it_elem in enumerate(it):
-        if it_elem == elem:
-            return i
-    return None
-
-
-def gcd(*args: sp.Monomial) -> sp.Monomial:
-    return sp.Monomial([min(m_i) for m_i in zip(*args)])
+def mlist_to_poly(mlist: Collection[sp.Monomial], gens) -> sp.Poly:
+    return sp.Poly(reduce(add, mlist), gens)
 
 
 def can_substitutions_quadratize(monom: sp.Monomial, subs: Iterable[sp.Monomial]) -> bool:
     gens = tuple(monom.gens)
     var_subs = set(map(lambda var: sp.Monomial(var, gens), gens))
     var_subs.add(sp.Monomial((0,) * len(gens), gens))  # zero degree monomial
-    # unified_gens_subs = set(map(lambda sub: unify_monom_to_gens(sub, gens), subs))
     expanded_subs = var_subs.union(subs)
     return _can_quad_0(monom) or _can_quad_2(monom, expanded_subs) or _can_quad_1(monom, expanded_subs)
 
@@ -101,4 +85,3 @@ def _can_quad_1(monom: sp.Monomial, subs: Iterable[sp.Monomial]) -> bool:
 
 def _can_quad_0(monom: sp.Monomial) -> bool:
     return monomial_deg(monom) == 0
-
